Added NPC max hit testing
This is kind of like a unit test, but not really. We can check what NPCs have a max hit of 1 which seems to be the default in our NPC def json, and if necessary, we can update the JSON to change the max hit for NPCs that should actually be hitting higher than 1.